+++ apps/snappyHexMesh/view/config/content/4_Refinement/RefinementProps/Boundary.qml
@@ -16,15 +16,21 @@ Column {
         label: "Type"
         width: parent.width
         path: "refinement:Boundary.type"
+//        model: ListModel {
+//            ListElement { text: qsTr("patch") }
+//            ListElement { text: qsTr("wall") }
+//            ListElement { text: qsTr("symmetry") }
+//            ListElement { text: qsTr("empty") }
+//            ListElement { text: qsTr("wedge") }
+//            ListElement { text: qsTr("cyclic") }
+//            ListElement { text: qsTr("cyclicAMI") }
+//        }
         model: ListModel {
-            ListElement { text: qsTr("patch") }
-            ListElement { text: qsTr("wall") }
-            ListElement { text: qsTr("symmetryPlane") }
-            ListElement { text: qsTr("symmetry") }
-            ListElement { text: qsTr("empty") }
-            ListElement { text: qsTr("wedge") }
-            ListElement { text: qsTr("cyclic") }
-            ListElement { text: qsTr("cyclicAMI") }
+            Component.onCompleted: {
+                for (var i = 0; i < app.boundaryTypesModel.length; i++) {
+                    append({'text': app.boundaryTypesModel[i]})
+                }
+            }
         }
     }
 }